Safari-611.1.10/LayoutTests/platform/ios-wk1/platform/ios-simulator/ios/fast/events/touch/textarea-touch-target-expected.txt

Tests that input elements can receive touch events.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS event.type is "touchstart"
PASS event.touches[0].target.id is "target"
PASS event.touches[0].target.nodeName is "TEXTAREA"
PASS successfullyParsed is true

TEST COMPLETE
